// src/graphql/generated/server/graphql-resolvers.ts
export type Maybe<T> = T | null;
// ...
export type RequireFields<T, K extends keyof T> = Omit<T, K> & {
	[P in K]-?: NonNullable<T[P]>;
};
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
	ID: { input: string; output: string };
	String: { input: string; output: string };
	Boolean: { input: boolean; output: boolean };
	Int: { input: number; output: number };
	Float: { input: number; output: number };
};

export type Assignment = {
	__typename?: "Assignment";
	createdAt: Scalars["String"]["output"];
	dueDate: Scalars["String"]["output"];
	id: Scalars["Int"]["output"];
	key: Scalars["String"]["output"];
	ptSessionId: Scalars["Int"]["output"];
	taskId: Scalars["Int"]["output"];
	taskType: Scalars["String"]["output"];
	updatedAt: Scalars["String"]["output"];
};

export type BodyCondition = {
	__typename?: "BodyCondition";
	armCircumference: Maybe<Scalars["Float"]["output"]>;
	avgRespirationRate: Maybe<Scalars["Float"]["output"]>;
	avgSleepTime: Maybe<Scalars["Float"]["output"]>;
	bmi: Maybe<Scalars["Float"]["output"]>;
	bmr: Maybe<Scalars["Float"]["output"]>;
	bodyFat: Maybe<Scalars["Float"]["output"]>;
	bust: Maybe<Scalars["Float"]["output"]>;
	buttockHeight: Maybe<Scalars["Float"]["output"]>;
	calfCircumference: Maybe<Scalars["Float"]["output"]>;
	clientId: Scalars["Int"]["output"];
	ffd: Maybe<Scalars["Float"]["output"]>;
	hip: Maybe<Scalars["Float"]["output"]>;
	hwd: Maybe<Scalars["Float"]["output"]>;
	id: Scalars["Int"]["output"];
	measuredAt: Scalars["String"]["output"];
	memo: Maybe<Scalars["String"]["output"]>;
	muscleMass: Maybe<Scalars["Float"]["output"]>;
	parasympathetic: Maybe<Scalars["Float"]["output"]>;
	skeletalMuscleRate: Maybe<Scalars["Float"]["output"]>;
	sympathetic: Maybe<Scalars["Float"]["output"]>;
	thighCircumference: Maybe<Scalars["Float"]["output"]>;
	underbust: Maybe<Scalars["Float"]["output"]>;
	waist: Maybe<Scalars["Float"]["output"]>;
	weight: Maybe<Scalars["Float"]["output"]>;
};

export type Client = {
	__typename?: "Client";
	id: Scalars["Int"]["output"];
	profile: Maybe<ClientProfile>;
	sessions: Array<PtSession>;
	user: User;
	userId: Scalars["String"]["output"];
};

export type ClientProfile = {
	__typename?: "ClientProfile";
	allowSnsPost: Scalars["String"]["output"];
	clientId: Scalars["String"]["output"];
	exerciseHistory: Scalars["String"]["output"];
	hobby: Scalars["String"]["output"];
	id: Scalars["Int"]["output"];
	occupation: Scalars["String"]["output"];
};
// ...
export type Mutation = {
	__typename?: "Mutation";
	createAssignment: Assignment;
	createBodyCondition: BodyCondition;
	createClient: Client;
	createClientWithProfile: Client;
	createCurriculumUnit: CurriculumUnit;
	createLearningMaterial: LearningMaterial;
	createMidtermHealthPurpose: MidtermHealthPurpose;
	createRequiredFunction: RequiredFunction;
	createSession: PtSession;
	createTrainer: Trainer;
	createTrainingMenu: TrainingMenu;
	createTreatmentMenu: TreatmentMenu;
	createUser: User;
};

export type MutationCreateAssignmentArgs = {
	input: CreateAssignmentInput;
};

export type MutationCreateBodyConditionArgs = {
	input: CreateBodyConditionInput;
};

export type MutationCreateClientArgs = {
	input: CreateClientInput;
};

export type MutationCreateClientWithProfileArgs = {
	input: CreateClientWithProfileInput;
};

export type MutationCreateCurriculumUnitArgs = {
	input: CreateCurriculumUnitInput;
};

export type MutationCreateLearningMaterialArgs = {
	input: CreateLearningMaterialInput;
};
// ...
export type PtSession = {
	__typename?: "PTSession";
	client: Client;
	clientId: Scalars["String"]["output"];
	createdAt: Scalars["String"]["output"];
	duration: Scalars["Int"]["output"];
	id: Scalars["Int"]["output"];
	items: Array<PtSessionItem>;
	notes: Maybe<Scalars["String"]["output"]>;
	scheduledAt: Scalars["String"]["output"];
	status: SessionStatus;
	trainer: Trainer;
	trainerId: Scalars["String"]["output"];
	updatedAt: Scalars["String"]["output"];
};
// ...
export type User = {
	__typename?: "User";
	active_flag: Scalars["Boolean"]["output"];
	birth_date: Scalars["String"]["output"];
	created_at: Scalars["String"]["output"];
	credentials: Maybe<UserCredentials>;
	first_name: Scalars["String"]["output"];
	first_name_kana: Scalars["String"]["output"];
	gender: Scalars["Int"]["output"];
	id: Scalars["Int"]["output"];
	key: Scalars["String"]["output"];
	kind: Scalars["Int"]["output"];
	last_name: Scalars["String"]["output"];
	last_name_kana: Scalars["String"]["output"];
	updated_at: Scalars["String"]["output"];
};
// ...
export type Resolver<
	TResult,
	TParent = Record<PropertyKey, never>,
	TContext = Record<PropertyKey, never>,
	TArgs = Record<PropertyKey, never>,
> =
	| ResolverFn<TResult, TParent, TContext, TArgs>
	| ResolverWithResolve<TResult, TParent, TContext, TArgs>;
// ...
export type ResolversTypes = {
	Assignment: ResolverTypeWrapper<Assignment>;
	BodyCondition: ResolverTypeWrapper<BodyCondition>;
	Boolean: ResolverTypeWrapper<Scalars["Boolean"]["output"]>;
	Client: ResolverTypeWrapper<Client>;
	ClientProfile: ResolverTypeWrapper<ClientProfile>;
	CreateAssignmentInput: CreateAssignmentInput;
	CreateBodyConditionInput: CreateBodyConditionInput;
	CreateClientInput: CreateClientInput;
	CreateClientWithProfileInput: CreateClientWithProfileInput;
	CreateCurriculumUnitInput: CreateCurriculumUnitInput;
	CreateLearningMaterialInput: CreateLearningMaterialInput;
	CreateMidtermHealthPurposeInput: CreateMidtermHealthPurposeInput;
	CreateRequiredFunctionInput: CreateRequiredFunctionInput;
	CreateSessionInput: CreateSessionInput;
	CreateTrainerInput: CreateTrainerInput;
	CreateTrainingMenuInput: CreateTrainingMenuInput;
	CreateTreatmentMenuInput: CreateTreatmentMenuInput;
	CreateUserInput: CreateUserInput;
	CurriculumUnit: ResolverTypeWrapper<CurriculumUnit>;
	Float: ResolverTypeWrapper<Scalars["Float"]["output"]>;
	Gender: Gender;
	Int: ResolverTypeWrapper<Scalars["Int"]["output"]>;
	LearningMaterial: ResolverTypeWrapper<LearningMaterial>;
	MidtermHealthPurpose: ResolverTypeWrapper<MidtermHealthPurpose>;
	Mutation: ResolverTypeWrapper<Record<PropertyKey, never>>;
	PTSession: ResolverTypeWrapper<PtSession>;
	PTSessionItem: ResolverTypeWrapper<PtSessionItem>;
	Query: ResolverTypeWrapper<Record<PropertyKey, never>>;
	RequiredFunction: ResolverTypeWrapper<RequiredFunction>;
	SessionStatus: SessionStatus;
	String: ResolverTypeWrapper<Scalars["String"]["output"]>;
	Trainer: ResolverTypeWrapper<Trainer>;
	TrainerProfile: ResolverTypeWrapper<TrainerProfile>;
	TrainingMenu: ResolverTypeWrapper<TrainingMenu>;
	TreatmentMenu: ResolverTypeWrapper<TreatmentMenu>;
	User: ResolverTypeWrapper<User>;
	UserCredentials: ResolverTypeWrapper<UserCredentials>;
	UserRole: UserRole;
};

/** Mapping between all available schema types and the resolvers parents */
export type ResolversParentTypes = {
	Assignment: Assignment;
	BodyCondition: BodyCondition;
	Boolean: Scalars["Boolean"]["output"];
	Client: Client;
	ClientProfile: ClientProfile;
	CreateAssignmentInput: CreateAssignmentInput;
	CreateBodyConditionInput: CreateBodyConditionInput;
	CreateClientInput: CreateClientInput;
	CreateClientWithProfileInput: CreateClientWithProfileInput;
	CreateCurriculumUnitInput: CreateCurriculumUnitInput;
	CreateLearningMaterialInput: CreateLearningMaterialInput;
	CreateMidtermHealthPurposeInput: CreateMidtermHealthPurposeInput;
	CreateRequiredFunctionInput: CreateRequiredFunctionInput;
	CreateSessionInput: CreateSessionInput;
	CreateTrainerInput: CreateTrainerInput;
	CreateTrainingMenuInput: CreateTrainingMenuInput;
	CreateTreatmentMenuInput: CreateTreatmentMenuInput;
	CreateUserInput: CreateUserInput;
	CurriculumUnit: CurriculumUnit;
	Float: Scalars["Float"]["output"];
	Int: Scalars["Int"]["output"];
	LearningMaterial: LearningMaterial;
	MidtermHealthPurpose: MidtermHealthPurpose;
	Mutation: Record<PropertyKey, never>;
	PTSession: PtSession;
	PTSessionItem: PtSessionItem;
	Query: Record<PropertyKey, never>;
	RequiredFunction: RequiredFunction;
	String: Scalars["String"]["output"];
	Trainer: Trainer;
	TrainerProfile: TrainerProfile;
	TrainingMenu: TrainingMenu;
	TreatmentMenu: TreatmentMenu;
	User: User;
	UserCredentials: UserCredentials;
};
// ...
export type MutationResolvers<
	ContextType = Context,
	ParentType extends
		ResolversParentTypes["Mutation"] = ResolversParentTypes["Mutation"],
> = {
	createAssignment: Resolver<
		ResolversTypes["Assignment"],
		ParentType,
		ContextType,
		RequireFields<MutationCreateAssignmentArgs, "input">
	>;
	createBodyCondition: Resolver<
		ResolversTypes["BodyCondition"],
		ParentType,
		ContextType,
		RequireFields<MutationCreateBodyConditionArgs, "input">
	>;
	createClient: Resolver<
		ResolversTypes["Client"],
		ParentType,
		ContextType,
		RequireFields<MutationCreateClientArgs, "input">
	>;
	createClientWithProfile: Resolver<
		ResolversTypes["Client"],
		ParentType,
		ContextType,
		RequireFields<MutationCreateClientWithProfileArgs, "input">
	>;
	createCurriculumUnit: Resolver<
		ResolversTypes["CurriculumUnit"],
		ParentType,
		ContextType,
		RequireFields<MutationCreateCurriculumUnitArgs, "input">
	>;
	createLearningMaterial: Resolver<
		ResolversTypes["LearningMaterial"],
		ParentType,
		ContextType,
		RequireFields<MutationCreateLearningMaterialArgs, "input">
	>;
	createMidtermHealthPurpose: Resolver<
		ResolversTypes["MidtermHealthPurpose"],
		ParentType,
		ContextType,
		RequireFields<MutationCreateMidtermHealthPurposeArgs, "input">
	>;
	createRequiredFunction: Resolver<
		ResolversTypes["RequiredFunction"],
		ParentType,
		ContextType,
		RequireFields<MutationCreateRequiredFunctionArgs, "input">
	>;
	createSession: Resolver<
		ResolversTypes["PTSession"],
		ParentType,
		ContextType,
		RequireFields<MutationCreateSessionArgs, "input">
	>;
	createTrainer: Resolver<
		ResolversTypes["Trainer"],
		ParentType,
		ContextType,
		RequireFields<MutationCreateTrainerArgs, "input">
	>;
	createTrainingMenu: Resolver<
		ResolversTypes["TrainingMenu"],
		ParentType,
		ContextType,
		RequireFields<MutationCreateTrainingMenuArgs, "input">
	>;
	createTreatmentMenu: Resolver<
		ResolversTypes["TreatmentMenu"],
		ParentType,
		ContextType,
		RequireFields<MutationCreateTreatmentMenuArgs, "input">
	>;
	createUser: Resolver<
		ResolversTypes["User"],
		ParentType,
		ContextType,
		RequireFields<MutationCreateUserArgs, "input">
	>;
};
// ...
export type QueryResolvers<
	ContextType = Context,
	ParentType extends
		ResolversParentTypes["Query"] = ResolversParentTypes["Query"],
> = {
	assignments: Resolver<
		Array<ResolversTypes["Assignment"]>,
		ParentType,
		ContextType,
		RequireFields<QueryAssignmentsArgs, "sessionId">
	>;
	bodyConditions: Resolver<
		Array<ResolversTypes["BodyCondition"]>,
		ParentType,
		ContextType,
		RequireFields<QueryBodyConditionsArgs, "clientId">
	>;
	clients: Resolver<
		Array<ResolversTypes["Client"]>,
		ParentType,
		ContextType,
		QueryClientsArgs
	>;
	curriculumUnits: Resolver<
		Array<ResolversTypes["CurriculumUnit"]>,
		ParentType,
		ContextType,
		QueryCurriculumUnitsArgs
	>;
	hello: Resolver<Maybe<ResolversTypes["String"]>, ParentType, ContextType>;
	learningMaterials: Resolver<
		Array<ResolversTypes["LearningMaterial"]>,
		ParentType,
		ContextType,
		QueryLearningMaterialsArgs
	>;
	midtermHealthPurposes: Resolver<
		Array<ResolversTypes["MidtermHealthPurpose"]>,
		ParentType,
		ContextType,
		RequireFields<QueryMidtermHealthPurposesArgs, "clientId">
	>;
	requiredFunctions: Resolver<
		Array<ResolversTypes["RequiredFunction"]>,
		ParentType,
		ContextType,
		RequireFields<QueryRequiredFunctionsArgs, "curriculumUnitId">
	>;
	sessions: Resolver<
		Array<ResolversTypes["PTSession"]>,
		ParentType,
		ContextType,
		QuerySessionsArgs
	>;
	trainers: Resolver<
		Array<ResolversTypes["Trainer"]>,
		ParentType,
		ContextType,
		QueryTrainersArgs
	>;
	trainingMenus: Resolver<
		Array<ResolversTypes["TrainingMenu"]>,
		ParentType,
		ContextType,
		QueryTrainingMenusArgs
	>;
	treatmentMenus: Resolver<
		Array<ResolversTypes["TreatmentMenu"]>,
		ParentType,
		ContextType,
		QueryTreatmentMenusArgs
	>;
	users: Resolver<
		Array<ResolversTypes["User"]>,
		ParentType,
		ContextType,
		QueryUsersArgs
	>;
};